Time: 0 ms (100.00%) | Memory: 8 MB (14.72%) - LeetSync

+++ 476-number-complement/number-complement.cpp
@@ -17,6 +17,6 @@ public:
            power=power*2;
            n=n/2;
        }
-       return (int)r;
+       return r;
     }
 };
